adding sw-config
